DataSets/requirements.txt
pandas>=2.0.0
numpy>=1.24.0
marker-pdf>=0.2.0
